Fixed ms conversion and draw now happens on data receipt
I have made the draw happen when data is received.  I think there is
actually an argument not to do this because the draws will happen in
quick succession which could result in blinking (I think).  It’s
probably not worrying about since we should switch this to a real text
GUI toolkit to address other issues anyway.